Can read + modify fat attributes

+++ proj/src/kernel/fat_fs_utils.cpp
@@ -736,10 +736,10 @@ void update_size_file_in_folder(char *filename_path, size_t offset, size_t origi
     if (folders_in_path.size() == 0) { //jsme v root slozce
         data_clust_fol = read_data_from_fat_fs(sectors_nums_data.at(cluster_num) - 31, 1); //-31; fce cte z dat sektoru
 
-        data_clust_fol.at(item_num_clust_rel * 32 + 28) = new_file_size_hex.at(0);
-        data_clust_fol.at(item_num_clust_rel * 32 + 29) = new_file_size_hex.at(1);
-        data_clust_fol.at(item_num_clust_rel * 32 + 30) = new_file_size_hex.at(2);
-        data_clust_fol.at(item_num_clust_rel * 32 + 31) = new_file_size_hex.at(3);
+        data_clust_fol.at(static_cast<int>(item_num_clust_rel * 32) + 28) = new_file_size_hex.at(0);
+        data_clust_fol.at(static_cast<int>(item_num_clust_rel * 32) + 29) = new_file_size_hex.at(1);
+        data_clust_fol.at(static_cast<int>(item_num_clust_rel * 32) + 30) = new_file_size_hex.at(2);
+        data_clust_fol.at(static_cast<int>(item_num_clust_rel * 32) + 31) = new_file_size_hex.at(3);
 
         std::vector<char> data_to_save;
         for (int i = 0; i < data_clust_fol.size(); i++) {
@@ -751,10 +751,10 @@ void update_size_file_in_folder(char *filename_path, size_t offset, size_t origi
     else {
         data_clust_fol = read_data_from_fat_fs(sectors_nums_data.at(cluster_num), 1); //fce cte z dat sektoru
 
-        data_clust_fol.at(item_num_clust_rel * 32 + 28) = new_file_size_hex.at(0);
-        data_clust_fol.at(item_num_clust_rel * 32 + 29) = new_file_size_hex.at(1);
-        data_clust_fol.at(item_num_clust_rel * 32 + 30) = new_file_size_hex.at(2);
-        data_clust_fol.at(item_num_clust_rel * 32 + 31) = new_file_size_hex.at(3);
+        data_clust_fol.at(static_cast<int>(item_num_clust_rel * 32) + 28) = new_file_size_hex.at(0);
+        data_clust_fol.at(static_cast<int>(item_num_clust_rel * 32) + 29) = new_file_size_hex.at(1);
+        data_clust_fol.at(static_cast<int>(item_num_clust_rel * 32) + 30) = new_file_size_hex.at(2);
+        data_clust_fol.at(static_cast<int>(item_num_clust_rel * 32) + 31) = new_file_size_hex.at(3);
 
         std::vector<char> data_to_save;
         for (int i = 0; i < data_clust_fol.size(); i++) {
@@ -1286,14 +1286,14 @@ int allocate_new_cluster(int start_cluster, std::vector<int>& fat_table_dec, std
         //na novy index priradim 4095 - znaci konec slozky / souboru - START
         std::vector<unsigned char> modified_bytes = convert_num_to_bytes_fat(free_index, first_fat_table_hex, 4095);
         first_fat_table_hex.at(static_cast<int>(static_cast<double>(free_index) * 1.5)) = modified_bytes.at(0); //oznacit cluster jako konecny v hex tabulce
-        first_fat_table_hex.at((static_cast<int>(static_cast<double>(free_index) * 1.5)) + 1) = modified_bytes.at(1);
+        first_fat_table_hex.at((static_cast<int>(static_cast<double>(free_index) * 1.5)) + static_cast <int>(1)) = modified_bytes.at(1);
         fat_table_dec.at(free_index) = 4095; //oznacit cluster jako konecny v dec tabulce
         //na novy index priradim 4095 - znaci konec slozky / souboru - KONEC
 
         //na puvodne posledni cluster souboru navazu novy - START
         modified_bytes = convert_num_to_bytes_fat(item_clusters.at(item_clusters.size() - 1), first_fat_table_hex, free_index);
         first_fat_table_hex.at(static_cast<int>(static_cast<double>(item_clusters.at(item_clusters.size() - 1)) * 1.5)) = modified_bytes.at(0); //oznacit cluster jako konecny v hex tabulce
-        first_fat_table_hex.at((static_cast<int>(static_cast<double>(item_clusters.at(item_clusters.size() - 1)) * 1.5)) + 1) = modified_bytes.at(1);
+        first_fat_table_hex.at((static_cast<int>(static_cast<double>(item_clusters.at(item_clusters.size() - 1)) * 1.5)) + static_cast <int>(1)) = modified_bytes.at(1);
         fat_table_dec.at(item_clusters.at(item_clusters.size() - 1)) = free_index; //oznacit cluster jako konecny v dec tabulce
         //na puvodne posledni cluster souboru navazu novy - KONEC
